Refactored package change: model.entity to model only
