+++ server/src/main/kotlin/com/ecotrade/server/user/model/User.kt
@@ -1,4 +1,4 @@
-package com.ecotrade.server.user.model.entity
+package com.ecotrade.server.user.model
 
 import jakarta.persistence.*
 import jakarta.validation.constraints.Email
