+++ server/src/main/kotlin/com/ecotrade/server/user/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package com.ecotrade.server.user.repository
 
-import com.ecotrade.server.user.model.entity.User
+import com.ecotrade.server.user.model.User
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.Optional
