+++ server/src/main/kotlin/com/ecotrade/server/user/service/UserService.kt
@@ -1,6 +1,6 @@
 package com.ecotrade.server.user.service
 
-import com.ecotrade.server.user.model.entity.User
+import com.ecotrade.server.user.model.User
 import com.ecotrade.server.user.repository.UserRepository
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
